Revises evaluation to remove excluded annotations in properties (#1)
Overhauls how evaluate keeps track of which views to score, removing an old hack that modified the AnaforaData objects, and replacing it with a new `ToSet` object which converts just the selected portion of annotations to nested tuples. As an added benefit, the hacky _OverlappingWrapper class was no longer necessary.